# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Chat Service API",
    "version": "1.0.0",
    "description": "API completa para el microservicio de chat con soporte para IA y conversaciones 1 a 1. Autenticación simplificada basada en userId sin JWT.",
    "contact": {
      "name": "Equipo de Desarrollo",
      "email": "dev@example.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3003",
      "description": "Servidor de desarrollo"
    },
    {
      "url": "https://api.example.com",
      "description": "Servidor de producción"
    }
  ],
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "null"
          },
          "message": {
            "type": "string",
            "description": "Mensaje de error"
          },
          "status": {
            "type": "string",
            "enum": [
              "error"
            ]
          },
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "description": "Código de error"
              },
              "details": {
                "type": "array",
                "items": {
                  "type": "string"
                },
                "description": "Detalles del error"
              }
            }
          }
        }
      },
      "PaginationMeta": {
        "type": "object",
        "properties": {
          "page": {
            "type": "integer",
            "description": "Página actual"
          },
          "limit": {
            "type": "integer",
            "description": "Elementos por página"
          },
          "total": {
            "type": "integer",
            "description": "Total de elementos"
          },
          "totalPages": {
            "type": "integer",
            "description": "Total de páginas"
          },
          "hasNext": {
            "type": "boolean",
            "description": "Si hay siguiente página"
          },
          "hasPrev": {
            "type": "boolean",
            "description": "Si hay página anterior"
          }
        }
      },
      "ChatMessage": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "ID único del mensaje"
          },
          "mensaje": {
            "type": "string",
            "description": "Contenido del mensaje"
          },
          "estado": {
            "type": "string",
            "enum": [
              "enviado",
              "entregado",
              "leido",
              "fallido"
            ],
            "description": "Estado del mensaje"
          },
          "fecha": {
            "type": "string",
            "format": "date-time",
            "description": "Fecha de envío"
          },
          "usuario_id": {
            "type": "string",
            "description": "ID del usuario que envió el mensaje"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "Fecha de creación"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "description": "Fecha de última actualización"
          },
          "is_ai_response": {
            "type": "boolean",
            "description": "Si es respuesta de IA"
          },
          "response_to_message_id": {
            "type": "string",
            "description": "ID del mensaje al que responde"
          },
          "conversation_id": {
            "type": "string",
            "description": "ID de la conversación (para mensajes privados)"
          },
          "recipient_id": {
            "type": "string",
            "description": "ID del destinatario (para mensajes privados)"
          }
        }
      },
...
      "SendMessageRequest": {
        "type": "object",
        "required": [
          "mensaje",
          "usuario_id"
        ],
        "properties": {
          "mensaje": {
            "type": "string",
            "minLength": 1,
            "maxLength": 5000,
            "description": "Contenido del mensaje"
          },
          "usuario_id": {
            "type": "string",
            "description": "ID del usuario que envía el mensaje"
          },
          "chat_estudiante_id": {
            "type": "string",
            "description": "ID del estudiante (opcional)"
          },
          "recipient_id": {
            "type": "string",
            "description": "ID del destinatario (para mensajes privados)"
          },
          "conversation_id": {
            "type": "string",
            "description": "ID de la conversación (para continuar conversación existente)"
          }
        }
      },
      "SendMessageResponse": {
        "type": "object",
        "properties": {
          "message": {
            "$ref": "#/components/schemas/ChatMessage"
          },
          "ai_response": {
            "$ref": "#/components/schemas/ChatMessage"
          }
        }
      },
...
      "ChatAttempt": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "ID único del intento"
          },
          "estudiante_id": {
            "type": "string",
            "description": "ID del estudiante"
          },
          "fecha_inicio": {
            "type": "string",
            "format": "date-time",
            "description": "Fecha de inicio del intento"
          },
          "fecha_fin": {
            "type": "string",
            "format": "date-time",
            "description": "Fecha de fin del intento"
          },
          "duracion_minutos": {
            "type": "number",
            "description": "Duración en minutos"
          },
          "mensajes_enviados": {
            "type": "integer",
            "description": "Número de mensajes enviados"
          },
          "estado": {
            "type": "string",
            "enum": [
              "activo",
              "finalizado",
              "pausado"
            ],
            "description": "Estado del intento"
          }
        }
      },
...
    },
    "securitySchemes": {
      "UserIdAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-User-ID",
        "description": "ID del usuario (autenticación simplificada sin JWT)"
      }
    }
  },
...
  "paths": {
...
    "/chat/message": {
      "post": {
        "summary": "Enviar mensaje de chat",
        "description": "Envía un mensaje de chat y obtiene respuesta automática de IA",
        "tags": [
          "Chat"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendMessageRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Mensaje enviado exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/SendMessageResponse"
                    },
                    "message": {
                      "type": "string",
                      "example": "Mensaje enviado exitosamente y respuesta de IA generada"
                    },
                    "status": {
                      "type": "string",
                      "example": "success"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Datos de entrada inválidos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Error interno del servidor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/chat/history/{estudiante_id}": {
      "get": {
        "summary": "Obtener historial de chat",
        "description": "Obtiene el historial completo de chat de un estudiante",
        "tags": [
          "Chat"
        ],
        "parameters": [
          {
            "name": "estudiante_id",
            "in": "path",
            "required": true,
            "description": "ID del estudiante",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "description": "Número de página",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "description": "Elementos por página",
            "schema": {
              "type": "integer",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Historial obtenido exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "messages": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/ChatMessage"
                          }
                        },
                        "attempts": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/ChatAttempt"
                          }
                        },
                        "pagination": {
                          "$ref": "#/components/schemas/PaginationMeta"
                        }
                      }
                    },
                    "message": {
                      "type": "string",
                      "example": "Historial de chat obtenido exitosamente"
                    },
                    "status": {
                      "type": "string",
                      "example": "success"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Estudiante no encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/chat/history/{estudiante_id}/messages": {
      "get": {
        "summary": "Obtener solo mensajes",
        "description": "Obtiene solo los mensajes del historial (sin intentos) - Más rápido",
        "tags": [
          "Chat"
        ],
        "parameters": [
          {
            "name": "estudiante_id",
            "in": "path",
            "required": true,
            "description": "ID del estudiante",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "description": "Número de página",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "description": "Elementos por página",
            "schema": {
              "type": "integer",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Mensajes obtenidos exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "messages": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/ChatMessage"
                          }
                        },
                        "pagination": {
                          "$ref": "#/components/schemas/PaginationMeta"
                        }
                      }
                    },
                    "message": {
                      "type": "string",
                      "example": "Mensajes obtenidos exitosamente"
                    },
                    "status": {
                      "type": "string",
                      "example": "success"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/chat/attempt": {
      "post": {
        "summary": "Registrar intento de chat",
        "description": "Registra un intento de chat cuando se abre el input sin enviar",
        "tags": [
          "Chat"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "estudiante_id"
                ],
                "properties": {
                  "estudiante_id": {
                    "type": "string",
                    "description": "ID del estudiante"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Intento registrado exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/ChatAttempt"
                    },
                    "message": {
                      "type": "string",
                      "example": "Intento de chat registrado exitosamente"
                    },
                    "status": {
                      "type": "string",
                      "example": "success"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/chat/attempts/{estudiante_id}": {
      "get": {
        "summary": "Obtener intentos de chat",
        "description": "Obtiene los intentos de chat de un estudiante",
        "tags": [
          "Chat"
        ],
        "parameters": [
          {
            "name": "estudiante_id",
            "in": "path",
            "required": true,
            "description": "ID del estudiante",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Intentos obtenidos exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/ChatAttempt"
                      }
                    },
                    "message": {
                      "type": "string",
                      "example": "Intentos de chat obtenidos exitosamente"
                    },
                    "status": {
                      "type": "string",
                      "example": "success"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/chat/ai/test": {
      "post": {
        "summary": "Probar el servicio de IA",
        "description": "Prueba el servicio de IA con un mensaje de prueba",
        "tags": [
          "Chat"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "mensaje"
                ],
                "properties": {
                  "mensaje": {
                    "type": "string",
                    "description": "Mensaje de prueba para enviar a la IA",
                    "example": "Hola, ¿cómo estás?"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Prueba de IA completada exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "input": {
                          "type": "string",
                          "description": "Mensaje de entrada"
                        },
                        "response": {
                          "type": "string",
                          "description": "Respuesta de la IA"
                        },
                        "model": {
                          "type": "string",
                          "example": "gemini-pro"
                        },
                        "tokensUsed": {
                          "type": "integer",
                          "description": "Tokens utilizados"
                        }
                      }
                    },
                    "message": {
                      "type": "string",
                      "example": "Prueba de IA completada exitosamente"
                    },
                    "status": {
                      "type": "string",
                      "example": "success"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Mensaje de prueba requerido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Error probando IA",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  }
}